Light source influence calculation wasn't called if progress bars weren't enabled; After loading a save game, weapon icons didn't show which mine type was armed; Calculation of maximum robot strength was wrong, leading to damage indicators not reflecting the actual amount of damage a robot had taken (particularly boss robots were affected); Ship wiggling cannot be turned off in multiplayer games any more (not that it would make a difference ...); Added progress bar for loading hires models; Multiplayer option for towing flags wasn't recognized; Potential crash in D2X multiplayer options menu; Multiplayer competition parameter disabling certain D2X-XL extensions for multiplayer games; D2X-XL specific options were still present in the more multiplayer options menu
git-svn-id: svn://svn.code.sf.net/p/d2x-xl/code/trunk@278 01821946-851b-0410-b7f4-f1477c70e57a

+++ cfile/cfile.c
@@ -965,7 +965,7 @@ return pData;
 
 void CFSplitPath (char *szFullPath, char *szFolder, char *szFile, char *szExt)
 {
-	int	i, j, l = strlen (szFullPath) - 1;
+	int	i, j, l = (int) strlen (szFullPath) - 1;
 
 i = l;
 #ifdef _WIN32
